+++ src/main/scala/com/atomist/rug/test/gherkin/handler/event/EventHandlerFeature.scala
@@ -1,7 +1,7 @@
 package com.atomist.rug.test.gherkin.handler.event
 
 import com.atomist.project.archive.Rugs
-import com.atomist.rug.test.gherkin.{AbstractExecutableFeature, Definitions, FeatureDefinition, GherkinExecutionListener}
+import com.atomist.rug.test.gherkin._
 import com.atomist.source.ArtifactSource
 
 class EventHandlerFeature(
@@ -9,10 +9,11 @@ class EventHandlerFeature(
                       definitions: Definitions,
                       rugArchive: ArtifactSource,
                       rugs: Option[Rugs],
-                      listeners: Seq[GherkinExecutionListener])
-  extends AbstractExecutableFeature[EventHandlerScenarioWorld](definition, definitions, rugs, listeners) {
+                      listeners: Seq[GherkinExecutionListener],
+                      config: GherkinRunnerConfig)
+  extends AbstractExecutableFeature[EventHandlerScenarioWorld](definition, definitions, rugs, listeners, config) {
 
   override protected def createWorldForScenario(): EventHandlerScenarioWorld = {
-    new EventHandlerScenarioWorld(definitions, rugs, listeners)
+    new EventHandlerScenarioWorld(definitions, rugs, listeners, config)
   }
 }