+++ src/main/scala/com/atomist/util/ServiceLoaderBackedExtensionProvider.scala
@@ -15,11 +15,13 @@ import scala.reflect.{ClassTag, _}
 class ServiceLoaderBackedExtensionProvider[T: ClassTag](val keyProvider: T => String)
   extends LazyLogging {
 
+  // The following can be cached as it creates issues in shared class loader hiarchies
   def providerMap: Map[String, T] = {
+    logger.debug(s"Loading providers of type ${classTag[T].runtimeClass.getName} and class loader ${Thread.currentThread().getContextClassLoader}")
     ServiceLoader.load(classTag[T].runtimeClass).asScala.map {
       case t: T =>
         val key = keyProvider.apply(t)
-        logger.info(s"Registered provider '$key' with class '${t.getClass}'")
+        logger.debug(s"Registered provider '$key' with class '${t.getClass}'")
         key -> t
       case wtf =>
         throw new RugRuntimeException("Extension", s"Provider class ${wtf.getClass} must implement ${classTag[T].runtimeClass.getName} interface", null)